requirements.txt
pytest
pytest-mock
polars
rich
pyarrow
numpy
pandas
xlsxwriter
oracledb
python-dotenv
PySide6
python-docx
docxtpl
pymupdf>=1.22.0
